requirements.txt
asgiref==3.9.1
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
Django==5.2.4
django-crispy-forms==2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
